+++ packages/Core/tests/04-query.ts
@@ -2223,4 +2223,38 @@ describe("Testing Other Features", () => {
             done(e);
         })
     });
+
+    it("Conform Rows query works on entire table without select table.", (done: MochaDone) => {
+
+        const nSQL = new nanoSQL();
+
+        nSQL.createDatabase({
+            id: "my-db",
+            mode: "TEMP",
+            tables: [
+                {
+                    name: "users",
+                    model: {
+                        "uuid:uuid": { pk: true },
+                        "name:string": {},
+                        "age:int": {}
+                    }
+                }
+            ]
+        })
+        .then(() => {
+            // insert test data
+            return nSQL.selectTable("users").query("upsert", [
+                {name: "John", age: 23},
+                {name: "Dane", age: 27}
+            ]).exec()
+        })
+        .then(() => {
+            return nSQL.selectTable("users").query("conform rows").exec();
+        }).then(() => {
+            done();
+        }).catch((e) => {
+            done(e);
+        })
+    });
 });